Split reconciliation table as a spearate component
